refactor: project structure 수정
