+++ src/main/java/com/traders/traders/module/strategy/controller/dto/response/FindStrategiesInfoResponseDto.java
@@ -1,0 +1,14 @@
+package com.traders.traders.module.strategy.controller.dto.response;
+
+import java.util.List;
+
+import com.traders.traders.module.strategy.domain.repository.dao.StrategyInfoDao;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class FindStrategiesInfoResponseDto {
+	private final List<StrategyInfoDao> strategiesInfo;
+}
