+++ src/main/java/com/traders/traders/module/strategy/domain/repository/StrategyQueryRepository.java
@@ -1,0 +1,10 @@
+package com.traders.traders.module.strategy.domain.repository;
+
+import java.util.List;
+import java.util.Optional;
+
+import com.traders.traders.module.strategy.domain.repository.dao.StrategyInfoDao;
+
+public interface StrategyQueryRepository {
+	Optional<List<StrategyInfoDao>> findStrategiesInfoDao();
+}
