--- src/main/java/com/traders/traders/module/trading/strategy/domain/TradingType.java
@@ -1,12 +1,0 @@
-package com.traders.traders.module.trading.strategy.domain;
-
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public enum TradingType {
-	LONG("매수"),
-	SHORT("매도");
-
-	private final String value;
-}
